+++ views/layouts/footer.php
@@ -27,28 +27,51 @@
 <!-- AdminLTE App -->
 <script src="<?= $URL; ?>public/js/lib/adminlte/adminlte.min.js"></script>
 
-<!-- DataTables y extensiones (IMPORTANTE: mantener este orden) -->
-<script src="<?= $URL; ?>public/js/plugins/datatables/jquery.dataTables.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/dataTables.bootstrap4.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/dataTables.responsive.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/responsive.bootstrap4.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/dataTables.buttons.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/buttons.bootstrap4.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/buttons.html5.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/buttons.print.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/datatables/buttons.colVis.min.js"></script>
-
-<!-- Bibliotecas para exportación -->
-<script src="<?= $URL; ?>public/js/plugins/utils/jszip.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/utils/pdfmake.min.js"></script>
-<script src="<?= $URL; ?>public/js/plugins/utils/vfs_fonts.js"></script>
-
-<!-- Select2 -->
-<script src="<?= $URL; ?>public/js/plugins/select2/select2.min.js"></script>
-
 <!-- Scripts principales de la aplicación -->
 <script src="<?= $URL; ?>public/js/core/common-utils.js"></script>
-<script src="<?= $URL; ?>public/js/core/common-datatable.js"></script>
+
+<!-- Plugins condicionales JS -->
+<?php
+$plugin_js = [
+    'datatables' => [
+        'plugins/datatables/jquery.dataTables.min.js',
+        'plugins/datatables/dataTables.bootstrap4.min.js',
+        'plugins/datatables/dataTables.responsive.min.js',
+        'plugins/datatables/responsive.bootstrap4.min.js',
+        'plugins/datatables/dataTables.buttons.min.js',
+        'plugins/datatables/buttons.bootstrap4.min.js',
+        'plugins/datatables/buttons.html5.min.js',
+        'plugins/datatables/buttons.print.min.js',
+        'plugins/datatables/buttons.colVis.min.js',
+        'core/common-datatable.js',
+    ],
+    'datatables-export' => [
+        'plugins/utils/jszip.min.js',
+        'plugins/utils/pdfmake.min.js',
+        'plugins/utils/vfs_fonts.js',
+    ],
+    'select2' => [
+        'plugins/select2/select2.min.js',
+    ],
+    'validate' => [
+        'plugins/validations/jquery.validate.min.js',
+        'plugins/validations/additional-methods.min.js',
+        'core/common-validate.js',
+    ],
+    'chart' => [
+        'plugins/chart/Chart.js',
+    ],
+];
+if (isset($plugins) && is_array($plugins)):
+    foreach ($plugins as $plugin):
+        if (isset($plugin_js[$plugin])):
+            foreach ($plugin_js[$plugin] as $js): ?>
+                <script src="<?= $URL; ?>public/js/<?= $js; ?>"></script>
+<?php endforeach;
+        endif;
+    endforeach;
+endif;
+?>
 
 <!-- Scripts específicos por módulo -->
 <?php if (isset($module_scripts) && is_array($module_scripts)): ?>